GEO-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
anthropic
tavily-python
mistralai
plotly
pandas
requests
pydantic
supabase
bcrypt
PyJWT
google-auth
selenium
beautifulsoup4
playwright
webdriver-manager
psycopg2-binary
openai
razorpay
google-genai
lxml
python-multipart
paid-python
opentelemetry-sdk
opentelemetry-exporter-otlp
openinference-instrumentation-anthropic
openinference-instrumentation-openai